requirements.txt
Flask
boto3
azure-storage-blob
azure-ai-textanalytics
azure-core